requirements.txt
Django==5.0.6
djangorestframework==3.15.2
psycopg2==2.9.9
django-cors-headers==4.4.0